moved themes.jsx into styles.jsx
