+++ client/src/components/profile/UserProfile.jsx
@@ -7,7 +7,7 @@ import Status from './features/Status.jsx';
 import Feed from './features/Feed.jsx';
 import Subscribe from './features/Subscribe.jsx';
 import Orders from './features/Orders.jsx';
-import useStyles from './themes.jsx';
+import useStyles from './styles.jsx';
 
 const UserProfile = () => {
   const classes = useStyles();
@@ -52,7 +52,7 @@ const UserProfile = () => {
             <Grid container spacing={3}>
               <Grid item xs={12}>
                 <Paper className={classes.subscribe}>
-                  {/* <Subscribe /> */}
+                  <Subscribe />
                 </Paper>
               </Grid>
               <Grid item xs={12}>
