+++ client/src/components/profile/features/Status.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box, CardHeader, CardContext, TextField, Button } from '@mui/material';
 
 const Status = () => (
-  <h1>updata status</h1>
+  <h1>update status</h1>
 );
 
 export default Status;
