+++ client/src/components/profile/features/Subscribe.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 // import {  } from '../styles.jsx';
 
 const Subscribe = () => (
-  <h1>Subscribe </h1>
+  <h1>subscribe</h1>
 );
 
 export default Subscribe;
