--- client/src/components/profile/themes.jsx
@@ -1,35 +1,0 @@
-import { makeStyles } from '@mui/styles';
-
-const useStyles = makeStyles({
-  page: {
-
-  },
-  pic: {
-    minHeight: '20vh',
-  },
-  nutrition: {
-    minHeight: '65vh',
-  },
-  status: {
-    minHeight: '15vh',
-  },
-  feed: {
-    minHeight: '70vh',
-  },
-  subscribe: {
-    minHeight: '5vh',
-  },
-  order: {
-    minHeight: '22.69vh',
-  },
-  card: {
-    minheight: '30vh',
-  },
-
-  avatar: {
-    minHeight: '20vh',
-    minWidth: '10vw',
-  },
-});
-
-export default useStyles;
